Rename new EM adapter to PooledEM
