+++ spec/spec_helper.rb
@@ -14,9 +14,12 @@ require 'restify'
 
 if ENV['ADAPTER']
   case ENV['ADAPTER'].to_s.downcase
-    when 'em-http-request'
+    when 'em'
       require 'restify/adapter/em'
       Restify.adapter = Restify::Adapter::EM.new
+    when 'em-pooled'
+      require 'restify/adapter/pooled_em'
+      Restify.adapter = Restify::Adapter::PooledEM.new
     when 'typhoeus'
       require 'restify/adapter/typhoeus'
       Restify.adapter = Restify::Adapter::Typhoeus.new
